requirements.txt
apache-sedona==1.2.1
attrs==23.1.0
backoff==2.2.1
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
duckdb==0.8.0
exceptiongroup==1.0.4
Fiona==1.8.22
func-timeout==4.3.5
geopandas==0.12.1
greenlet==2.0.1
idna==3.4
importlib-resources==5.12.0
iniconfig==1.1.1
isort==5.12.0
jsonschema==4.17.3
more-itertools==9.1.0
munch==2.5.0
mypy-extensions==1.0.0
numpy==1.24.3
packaging==23.1
pandas==2.0.2
pathspec==0.11.1
pgcopy==1.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.5.3
pluggy==1.0.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
py4j==0.10.9
pyarrow==12.0.0
pyparsing==3.0.9
pyproj==3.4.0
pyrsistent==0.19.3
pyspark==3.0.3
pytest==7.2.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
Shapely==1.8.5.post1
six==1.16.0
SQLAlchemy==1.4.44
tomli==2.0.1
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.16
zipp==3.15.0
